Explicitly specify types to arguments of 'libc::syscall' (#74)
The 'libc::syscall' function uses varargs - as a result, its arguments
are completely untyped. THe user must ensure that it is called with the
proper types for the targeted syscall - otherwise, the calling
convention might cause arguments to be put into the wrong registers.

This commit explicitly casts the arguments to 'libc::syscall' to the
proper type for the 'getrandom' syscall. This ensures that the correct
types for the target platform will always be used, instead of relying on
the types used happening to match those required by the target platform.

This particular commit is a backport of
6716ad06a5a1e502b80bd1983308eab655d90b83, with the addition of
`sys_fill_exact` from master (originally committed in
65660e00bed828c577d01dfd1f28f3f89aa72c56) to make the backport more
obviously correct.